Selecting cards now kinda works during tricks but not pass rounds
A bit more refactoring had to occur to get this to work. Each
CardGrouping now takes a game_controller, rather than a game, so that
the classes have a way of affecting the game, and accessing its state.

Besides that,
interface functions were added to the CardGrouping class:
 - cardSelected
 - mouseOverCard
 - mouseOffCard
These each get triggered for various input (mouse) events. Hand then
implements those functions to make them do something for when a card is
in a Hand... but only then. An isCardValid function was added to Hand as
well, which is used by the handlers to determine if the card can be
selected. It is not yet fully functional though.