The game board now gets created before registering event handlers
Some events were failing because they were relying on GUI elements
having been created already. Now the GUI is created before any events
can be triggered. The player name creation was relying on the player_pos
to have been sent from the server, so that was moved outside of board
creation to the "enter_room" event.