Fixing bug where tween finishes getting tweened to a removed group
When a player discards a card, a tween starts sending the card to the
discard pile. When each round ends, the previous discard pile is
removed. If a tween was still occurring, and the discard pile gets
removed to start the next round, an exception would occur and the game
would stop. This is because there is logic for the tween to "snap" to
its final position, making use of the position of it's child and parent
- which were then undefined. I simply added an undefined check for this
case, which seems to work as expected.