Passing cards now works on the first round
This actually took quite a bit of changes. Let's start on the front end.
As you might imagine, the card input handlers were changed a bit. These
changes aren't great, but are a step in the right direction. There are a
decent amount of things that still don't work, like deselecting cards if
you decide you don't want to pass one. Also, the hit boxes for the
sprites are definitely off.

There was some information that wasn't being sent that wasn't needed to
make the front end work properly, so a decent amount of event handler
code got changed on both ends. There is a new event being sent:
"pass_round_started". This contains the id of the round, which is
necessary so when the pass cards are selected, we can tell the server
which pass round to update, the number of cards which are to be passed,
and the new phase. Trick turn was also updated with the new phase so
that phase no longer needs to be sent as a separate event. This seemed
like a better decision, since all the necessary information is being
sent at the same time now. This also changed the "enter_room" event for
the same reason.

The rest of the changes were mostly refactoring to make the code a
little nicer, or were updates for the new events to be used.