Fixing exception which occurred on the second pass round
The bug was due to the following:
After the last trick discard was made, a delay message was sent.
However, the game, at that point, had already moved on to the next round
and marked the previous round as inactive. So when the game queried for
an active trick turn, the query failed and an exception was thrown. I
moved the query and put a try catch around it. Now, it is assumed if the
query fails, the move is just invalid.

Also, I found another bug where the selected_cards list was not getting
emptied after pass rounds, so that happens now.

Also, an event update wasn't captured in previous commits, so that is
being committed with this.