Fixing timer
There were a few issues with the timer, some on the front end and some
on the back end.

On the band end, the workers were not saving the time each move took
place, so the server didn't know how long the user actually took to make
each move. bank_ms is now properly saved. A few changes to the events
needed to be made.

On the front end, mostly the changes were to handle the new events.
Also, the timer now displays the amount of time the user actually has,
which was broken in previous versions.

I created an EnumField class to make it easier to use enums without
having to do a conversion to and from every time. PASS_DIRECTION is now
an enum being handled properly with ORM objects.

Dummy objects were created to help distinguish between real players and
bots so that delays could be adjusted. It's really to make it easier on
the developer, so that you doesn't have to wait the full amount of time
to continue, while still having enough time to make your own move.
PlayerType was created as an enum now using the new EnumField as well.

The length of the Player.channel field was increased because the channel
name isn't always the same. I think it has to do with the type of
service handling the channel (e.g. daphne). Hopefully, we don't get
longer names in the future. We might have to find a better long term
solution.

Unit tests were added. I would like to really focus on adding unit tests
in the future, which is something that has up till now been neglected.

Some with.transaction.atomic() blocks were added to avoid possible
synchronization issues.